Apply Marpit SVG polyfill for WebKit in Marp.ready()
