+++ package.json
@@ -79,7 +79,8 @@
     "typescript": "^3.2.2"
   },
   "dependencies": {
-    "@marp-team/marpit": "^0.4.1",
+    "@marp-team/marpit": "^0.5.0",
+    "@marp-team/marpit-svg-polyfill": "^0.1.0",
     "emoji-regex": "^7.0.3",
     "highlight.js": "^9.13.1",
     "katex": "^0.10.0",
